Fix absolute paths in exported CMake config file.
The current CONFIGURE_FILE() command produces a libtinsConfig.cmake
file with a hardcoded absolute path to the include directory.
As a result, the generated config file is not relocatable.

This is especially difficult when cross compiling where it is often
desirable to have a directory that mirrors the root filesystem of
your target (CMAKE_FIND_ROOT_PATH).  Since the config file has
absolute paths, it can't easily be shared amongst team members or
committed to version control.

This commit uses CONFIGURE_PACKAGE_CONFIG_FILE() instead, which
avoids hardcoded paths in the generated config file.

For each PATH_VAR passed in, CMake sets a PACKAGE_<var> variable
that is relative to the installed location of the package.

See http://www.cmake.org/cmake/help/v3.0/module/CMakePackageConfigHelpers.html